use uint32/64 to guarantee # of bytes and type cast shifts

+++ hw4/include/custom_functions.h
@@ -23,7 +23,7 @@ int printer_count;
  */
 typedef struct job {
     char *file;
-    int eligible; // 32 bits. One for each printer (1 if eligible)
+    uint32_t eligible; // 32 bits. One for each printer (1 if eligible)
     FILE_TYPE *type;
     JOB_STATUS status;
 } JOB;
@@ -33,7 +33,7 @@ typedef struct job {
  */
 JOB jobs[MAX_JOBS];
 // Global counter to keep track of the number of jobs. (64 bits - if bit is 1 then it is taken)
-size_t job_count;
+uint64_t job_count;
 
 /*
  * Space to store the PRINTER/JOB group pids.
